requirements.txt
numpy>=1.21.0
torch>=1.9.0
matplotlib>=3.4.0
pyyaml>=5.4.0
tqdm>=4.61.0
tensorboard>=2.6.0
scikit-learn>=0.24.0
pandas>=1.3.0
seaborn>=0.11.0
rich>=10.0.0
pytest>=6.2.0  # 用于测试
black>=21.5b2  # 用于代码格式化
flake8>=3.9.0  # 用于代码检查
mypy>=0.910    # 用于类型检查 